Add files via upload
Added the End Turn button to create the enemy players turn. Fixed code so that the enemy can attack the player. Need to change some values to make critical strikes work properly.
Must create code that will allow the javascript code to work on a mobile device. This will allow people to play the game on both mobile phone and PC